Add files via upload
Twitter sentiment analysis using machine learning (ML) is the process of analyzing tweets to determine the sentiment expressed within them, whether it's positive, negative, or neutral. This involves collecting tweets related to a particular topic or keyword, preprocessing the text data, and then using machine learning algorithms to classify the sentiment of each tweet. The aim is to understand public opinion or attitudes towards a specific subject, brand, or event by analyzing the sentiments expressed in tweets. This analysis can be valuable for businesses, organizations, and individuals to make informed decisions, monitor brand reputation, or gauge public perception in real-time.